Added router in main index.mjs
